feat: 通过了busybox的测例

+++ ulib/axstarry/src/trap.rs
@@ -1,5 +1,7 @@
 use axhal::{arch::TrapFrame, mem::VirtAddr, paging::MappingFlags};
 
+use crate::syscall::syscall;
+
 struct TrapHandlerImpl;
 
 #[crate_interface::impl_interface]
@@ -9,47 +11,44 @@ impl axhal::trap::TrapHandler for TrapHandlerImpl {
         {
             let guard = kernel_guard::NoPreempt::new();
             // trap进来，统计时间信息
-            // axprocess::time_stat_from_user_to_kernel();
+            axprocess::time_stat_from_user_to_kernel();
             axhal::irq::dispatch_irq(_irq_num);
-            // axprocess::time_stat_from_kernel_to_user();
+            axprocess::time_stat_from_kernel_to_user();
             drop(guard); // rescheduling may occur when preemption is re-enabled.
         }
     }
-    fn handle_syscall(_syscall_id: usize, _args: [usize; 6]) -> isize {
-        // axprocess::time_stat_from_user_to_kernel();
-        // let ans = syscall(syscall_id, args);
-        // axprocess::time_stat_from_kernel_to_user();
-        // ans
-        panic!();
-        // 0
+
+    fn handle_syscall(syscall_id: usize, args: [usize; 6]) -> isize {
+        axprocess::time_stat_from_user_to_kernel();
+        let ans = syscall(syscall_id, args);
+        axprocess::time_stat_from_kernel_to_user();
+        ans
     }
 
     #[cfg(feature = "paging")]
-    fn handle_page_fault(addr: VirtAddr, flags: MappingFlags, _tf: &mut TrapFrame) {
-        use axlog::info;
+    fn handle_page_fault(addr: VirtAddr, flags: MappingFlags, tf: &mut TrapFrame) {
+        use axhal::arch::SIGNAL_RETURN_TRAP;
         use axprocess::handle_page_fault;
-        // use axsignal::signal_no::SignalNo;
+        use axsignal::signal_no::SignalNo;
         use axtask::current;
         axprocess::time_stat_from_user_to_kernel();
-        // use crate::syscall::signal::{syscall_sigreturn, syscall_tkill};
-        // if addr.as_usize() == SIGNAL_RETURN_TRAP {
-        //     // 说明是信号执行完毕，此时应当执行sig return
-        //     tf.regs.a0 = syscall_sigreturn() as usize;
-        //     return;
-        // }
-        info!("in");
+        use crate::syscall::signal::{syscall_sigreturn, syscall_tkill};
+        if addr.as_usize() == SIGNAL_RETURN_TRAP {
+            // 说明是信号执行完毕，此时应当执行sig return
+            tf.regs.a0 = syscall_sigreturn() as usize;
+            return;
+        }
         if handle_page_fault(addr, flags).is_err() {
             // 如果处理失败，则发出sigsegv信号
             let curr = current().id().as_u64() as isize;
             axlog::error!("kill task: {}", curr);
-            // syscall_tkill(curr, SignalNo::SIGSEGV as isize);
+            syscall_tkill(curr, SignalNo::SIGSEGV as isize);
         }
-        info!("out");
         axprocess::time_stat_from_kernel_to_user();
     }
 
-    // #[cfg(feature = "signal")]
-    // fn handle_signal() {
-    //     axprocess::handle_signals();
-    // }
+    #[cfg(feature = "signal")]
+    fn handle_signal() {
+        axprocess::signal::handle_signals();
+    }
 }
